requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
